omap3: Fix voltage and frequency scaling order when changing OPP via mpurate
When lowering an OPP via mpurate from what is being set by bootloader,
the frequency need to be lowered before voltage is ramped down. The
current code was not taking this into consideration and was always
adjusting the voltage before the frequency adjustments.

This was leading to crashes on 3730 when mpurate was set to 300MHz
because the voltage was lowered before the frequency.

This patch fixes these issues by tracking the direction of OPP change
and doing the voltage and frequency change in the right order.

Signed-off-by: Ranjith Lohithakshan <ranjithl@ti.com>