mt9t112: Remove "smart" size selection
This code is broken, as it doesn't really adjust all other
settings, but output size to desired size. It isn't really
toggling everything that needs, and giving the false impression
of good size flexibility.

So, meanwhile, let's put fixed init settings for the contexts,
and toggle between them (implemented in other patch).

Signed-off-by: Sergio Aguirre <saaguirre@ti.com>