OMAP3EVM: Set minimum throughput requirement for DSS
GFX_FIFO_UNDERFLOW has been observed when the L3 rate is lowered
below 100MHz or so. Once this underflow happens the display stops
working.

This patch puts a minimum bus throughput requirement when lcd, tv
or dvi displays are enabled. The throughput is calculated in such
a manner that it will translate into a minimum L3 rate of 100MHz.

Signed-off-by: Ranjith Lohithakshan <ranjithl@ti.com>